Save image and generate html
Initial implementation of creating am html page for each comic image. Currently traversing backwards from most recent page.

I'll need to rename stuff and update the readme to reflect the change in direction, but it seems to work well.

It will overwrite the first existing html page it finds. This allows it to update the 'next' link on that page, which will otherwise just read 'current'.