Avoid Looping
Script now checks current page against secondMostRecent page to confirm it isn't bouncing back and forth between the last pages/first of a comic. This can be an issue if the next/previous buttons switch locations on the first or last pages of a comic.

This further supports the ability to start from the most recent page and work back.

Minor change to readme for clarity