Made image reference dynamic
Made the img src in the html file dynamic based on the name of the html file itself. As long as they match, the image will show up. The goal would be to allow renaming the files as needed. Unfortunately the 'next' button linking will fail if you do at the moment. But this little change helps some.